# [WIP] Rational Riccati solver