--- a/sympy/solvers/ode/ode.py
+++ b/sympy/solvers/ode/ode.py
@@ -294,6 +294,7 @@ allhints = (
     "1st_exact",
     "1st_linear",
     "Bernoulli",
+    "1st_rational_riccati",
     "Riccati_special_minus2",
     "1st_homogeneous_coeff_best",
     "1st_homogeneous_coeff_subs_indep_div_dep",
@@ -332,7 +333,6 @@ allhints = (
     "Liouville_Integral",
     "2nd_nonlinear_autonomous_conserved",
     "2nd_nonlinear_autonomous_conserved_Integral",
-    "1st_rational_riccati",
     )
 
 lie_heuristics = (