… doc/src/modules/solvers/ode.rst
… sympy/solvers/ode/single.py
… sympy/solvers/ode/tests/test_riccati.py
… sympy/solvers/ode/tests/test_single.py