This PR introduces a Rational Riccati ODE solver. This code can also be used to solve a subclass of 2nd order ODEs solvable by [Kovacic's Algorithm](https://cs.uwaterloo.ca/research/tr/1984/CS-84-35.pdf)

#### References to other Issues or PRs
Partial Fix #19183, #21503

#### Brief description of what is fixed or changed
The algorithm present [here](https://www3.risc.jku.at/publications/download/risc_5387/PhDThesisThieu.pdf) is implemented. The examples used to test are [here](https://www3.risc.jku.at/publications/download/risc_5197/RISCReport15-19.pdf).

#### Other comments
I will soon add the Kovacic solver.

#### Release Notes
<!-- BEGIN RELEASE NOTES -->
* solvers
  * ode:  Added a new solver for Riccati equation
<!-- END RELEASE NOTES -->