@oscarbenjamin 2021-05-11
I haven't looked through this in detail but it's good to see this coming.

I would put the algorithm in a separate file like `ricatti.py`. There can be a solver in `single.py` but it should just be a wrapper around the basic algorithm routines from `ricatti.py`. The basic routines should each be tested separately from dsolve.

The code needs to give much more explanation about the algorithm. Imagine someone in future who hasn't read the papers on which the algorithm is based trying to fix a simple bug. It's too much for them to go read a PhD thesis to understand what your code is doing. The top of the ricatti.py module should outline the algorithm and should also make it very clear what the reference for the algorithm is. Any assumptions you make and anything that you are not 100% sure about should be clearly documented within the code so that someone in future can understand what you were thinking when you wrote that code (e.g. if the assumption turns out to be incorrect in some cases).