@oscarbenjamin 2021-05-23
> Is the current structure okay?

The basic structure looks okay.

I'm going to hold off on detailed comments for now because there is one big question that needs to be resolved which is whether this algorithm should be formulated in terms of Poly rather than Expr.

There are lots of places which are using cancel, together, `as_numer_denom` etc which suggests that you should really be using a pair of Poly rather than an Expr. Every time you call cancel the whole expression is converted into a pair of Poly and then the `Poly.cancel` method is used and then the results are converted back to Expr. This is very inefficient compared to just working with `Poly` and using the `Poly.cancel` method.

Poly expressions are always in a canonical form which simplifies a lot of things and makes algebra more robust, predictable and efficient for an algorithm such as this.

Currently this uses `series` to compute Laurent series but the series function is slow because it is intended for more general cases. I'm pretty sure that the relevant Laurent coefficients for a rational function can be computed more efficiently using only polynomial operations.

If this can be written in terms of Poly and if it is better to do so then it is better to make that decision now as it will only be harder later. This is especially important if this is just the first of many related algorithms that should all share code and use the same kind of internal representation.

There are certain limitations of Poly such as the fact that symbolic powers of the independent variable would not be possible e.g. `x/(x + 1)**n` where `n` is a symbol. If this algorithm also has the same kind of limitations then I think it would be better to use Poly.

We can discuss this when we meet.