@naveensaigit 2021-05-24
I think using `Poly` would be best since finding the degree is not possible for polynomials with symbolic powers (which means the algorithm won't work for equations with symbolic powers in x).

As for Laurent series, I'll have to see how to get the coefficients in an efficient way without using the inbuilt function. I think that there is something mentioned in Smith's implementation of Kovavic's algorithm about finding the coefficients.