@asmeurer 2021-05-27
If we use Poly, these helper functions in risch will be useful. https://github.com/sympy/sympy/blob/31d469a5335c81ec4a437e36a861945a6b43d916/sympy/integrals/risch.py#L811-L886