@oscarbenjamin 2021-05-27
To make the substitution `x -> 1/x` in a rational function represented as two Polys you can use `num.transform`:
```python
In [11]: f = (x**2 - 1)/(x**2 + 1)

In [12]: f
Out[12]: 
 2    
x  - 1
──────
 2    
x  + 1

In [13]: num, den = [Poly(e, x) for e in f.as_numer_denom()]

In [14]: num
Out[14]: Poly(x**2 - 1, x, domain='ZZ')

In [15]: den
Out[15]: Poly(x**2 + 1, x, domain='ZZ')

In [18]: num.transform(Poly(1, x), Poly(x, x))
Out[18]: Poly(-x**2 + 1, x, domain='ZZ')

In [19]: den.transform(Poly(1, x), Poly(x, x))
Out[19]: Poly(x**2 + 1, x, domain='ZZ')

In [20]: num.transform(Poly(1, x), Poly(x, x)) / den.transform(Poly(1, x), Poly(x, x))
Out[20]: 
     2
1 - x 
──────
 2    
x  + 1
```
If the degree of the numerator and denominator are different then some factors of `x` need to be added somewhere.