@naveensaigit 2021-05-28
> If the degree of the numerator and denominator are different then some factors of `x` need to be added somewhere.

I think this should work.
``` python
In [3]: f = (x**2 - 1)/x**3

In [4]: f.subs(x, 1/x)
Out[4]: 
 3 ⎛     1 ⎞
x ⋅⎜-1 + ──⎟
   ⎜      2⎟
   ⎝     x ⎠

In [5]: num, den = [Poly(e, x) for e in f.as_numer_denom()]

In [6]: num.transform(Poly(1, x), Poly(x, x)) / den.transform(Poly(1, x), Poly(x, x))
Out[6]: 
     2
1 - x

In [7]: (num.transform(Poly(1, x), Poly(x, x)) / den.transform(Poly(1, x), Poly(x, x)))*x**(den.degree(x) - num.degree(x))
Out[7]: 
  ⎛     2⎞
x⋅⎝1 - x ⎠
```