@naveensaigit 2021-05-28
We were discussing yesterday that it's not possible to use `series` with `Poly`. Should we use `Expr` for this, or should this be implemented?