@oscarbenjamin 2021-05-31
> Actually, I am checking for `b2 != 0`, and whether `b0, b1 and b2` are rational functions in `_verify` before passing these to `solve_riccati`. The case where `b2 == 0` would become a Bernoulli equation (`y' = b0 + b1*y`), so it can already be solved by `dsolve`.

Think of this as something that should work on its own separately from the rest of `dsolve`. We should make a complete set of functions for working with and solving Riccatti ODEs. This should be usable independently of `dsolve` and it should be tested independently of `dsolve`. The goal here is not just to make it so that `dsolve` can find the solutions of this particular type of ODE but to build the first part in a series of interdependent algorithms. Of course we also want to add the capability to solve this kind of equation to `dsolve` but actually the main reason for implementing this algorithm is because it is the first steps for a better Lie group solver and for Kovacic's algorithm. In that wider context I am not sure if the case `b2 == 0` also needs to be handled.

> I am using `eq` to get the constant.

The constant can just be `Dummy('C')`.

> I'm guessing it is because of `expand` because I am substituting a big expression into the auxiliary differential equation and then expanding it out while solving it.

Okay let's follow this through for an example. Take this ODE and solution:
```python
In [16]: eq = nsimplify(Eq(Derivative(y(x), x), (1/3 - x)*y(x)**2/(-3*x - 1) + (x/2 - 1/3)*y(x)/(x - 3) + (11*x**3 + 323*x**2 - 315*x - 159)/(36*x**4 - 168*x*
    ...: *3 + 192*x**2 - 24*x - 36) - 1/3))

In [17]: sol = nsimplify(Eq(y(x), (2*x + 3/2)/(1 - x)))

In [18]: eq
Out[18]: 
                             ⎛x   1⎞                                               
                      2      ⎜─ - ─⎟⋅y(x)           3        2                     
d          (1/3 - x)⋅y (x)   ⎝2   3⎠            11⋅x  + 323⋅x  - 315⋅x - 159      1
──(y(x)) = ─────────────── + ──────────── + ─────────────────────────────────── - ─
dx             -3⋅x - 1         x - 3           4        3        2               3
                                            36⋅x  - 168⋅x  + 192⋅x  - 24⋅x - 36    

In [19]: sol
Out[19]: 
       2⋅x + 3/2
y(x) = ─────────
         1 - x 

In [20]: checkodesol(eq, sol)
Out[20]: (True, 0)
```
Now when we get to the `solve_aux_eq` function it is called with the arguments:
```python
(Pdb) p a
(1/3 - x)*(-(11*x**3 + 323*x**2 - 315*x - 159)/(36*x**4 - 168*x**3 + 192*x**2 - 24*x - 36) + 1/3)/(-3*x - 1) + (x/2 - 1/3)**2/(4*(x - 3)**2) + (x/2 - 1/3)/(2*(x - 3)**2) - 1/(4*(x - 3)) + (-3*x - 1)*(x/2 - 1/3)*(3*(1/3 - x)/(-3*x - 1)**2 - 1/(-3*x - 1))/(2*(1/3 - x)*(x - 3)) - 3*(-3*x - 1)*((3*x - 1)/(3*x + 1) - 1)/((1/3 - x)*(3*x + 1)**2) + 3*(-3*x - 1)**2*(3*(1/3 - x)/(-3*x - 1)**2 - 1/(-3*x - 1))**2/(4*(1/3 - x)**2)
(Pdb) p x
x
(Pdb) p m
0
(Pdb) p ybar
5/12 + 23/(36*(x + 1/3)) - 1/(2*(x - 1/3)) + 7/(12*(x - 1)) - 7/(12*(x - 3))
```
The function then does:
```python
auxeq = numer((p(x).diff(x, 2) + 2*ybar*p(x).diff(x) + (ybar.diff(x) + ybar**2 - a)*p(x)).together())
```
That gives:
```python
(Pdb) p auxeq
144*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)**2*(x - 1)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3)*Derivative(p(x), (x, 2)) + 24*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)*(x - 1)*(3*x - 1)*(3*x + 1)**2*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3)*(5*(x - 3)*(x - 1)*(3*x - 1)*(3*x + 1) + 23*(x - 3)*(x - 1)*(3*x - 1) - 18*(x - 3)*(x - 1)*(3*x + 1) + 7*(x - 3)*(3*x - 1)*(3*x + 1) - 7*(x - 1)*(3*x - 1)*(3*x + 1))*Derivative(p(x), x) + (-4*(1 - 3*x)**3*(-3*x - 1)*(x - 3)**2*(x - 1)**2*(3*x - 1)**2*(3*x + 1)**3*(12*x**4 - 67*x**3 - 259*x**2 + 307*x + 147) - 828*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)**2*(x - 1)**2*(3*x - 1)**2*(3*x + 1)*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) + 648*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)**2*(x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - 84*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) + 36*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)*(x - 1)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - (1 - 3*x)**2*(-3*x - 1)**2*(x - 1)**2*(3*x - 2)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - 12*(1 - 3*x)**2*(-3*x - 1)**2*(x - 1)**2*(3*x - 2)*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) + 84*(1 - 3*x)**2*(-3*x - 1)**2*(x - 1)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) + (1 - 3*x)**2*(-3*x - 1)**2*(3*x + 1)*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3)*(5*(x - 3)*(x - 1)*(3*x - 1)*(3*x + 1) + 23*(x - 3)*(x - 1)*(3*x - 1) - 18*(x - 3)*(x - 1)*(3*x + 1) + 7*(x - 3)*(3*x - 1)*(3*x + 1) - 7*(x - 1)*(3*x - 1)*(3*x + 1))**2 - 2592*(1 - 3*x)*(-3*x - 1)**3*(x - 3)**2*(x - 1)**2*(3*x - 1)**2*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - 72*(1 - 3*x)*(-3*x - 1)*(x - 3)*(x - 1)**2*(3*x - 2)*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - 3888*(x - 3)**2*(x - 1)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3))*p(x)
```
Pretty printed that is:
```python
                                                                                                      2                                                     
             2           2        2        2          2          3 ⎛   4       3       2          ⎞  d                      2           2                   
144⋅(1 - 3⋅x) ⋅(-3⋅x - 1) ⋅(x - 3) ⋅(x - 1) ⋅(3⋅x - 1) ⋅(3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠⋅───(p(x)) + 24⋅(1 - 3⋅x) ⋅(-3⋅x - 1) ⋅(x - 3)⋅(x - 1)⋅(3
                                                                                                      2                                                     
                                                                                                    dx                                                      

                                                                                                                                                            
                 2 ⎛   4       3       2          ⎞                                                                                                         
⋅x - 1)⋅(3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠⋅((x - 1)⋅(3⋅x - 1)⋅(3⋅x + 1)⋅(5⋅x - 15) + (x - 1)⋅(3⋅x - 1)⋅(23⋅x - 69) - (x - 1)⋅(3⋅x + 1)⋅(18⋅x - 54) 
                                                                                                                                                            
                                                                                                                                                            

                                                                                                                                                            
                                                                  d          ⎛             3                   2        2          2          3 ⎛    4      
+ (3⋅x - 1)⋅(3⋅x + 1)⋅(7⋅x - 21) - (3⋅x - 1)⋅(3⋅x + 1)⋅(7⋅x - 7))⋅──(p(x)) + ⎝- 4⋅(1 - 3⋅x) ⋅(-3⋅x - 1)⋅(x - 3) ⋅(x - 1) ⋅(3⋅x - 1) ⋅(3⋅x + 1) ⋅⎝12⋅x  - 67⋅
                                                                  dx                                                                                        
                                                                                                                                                            

                                                                                                                                                            
 3        2              ⎞                2           2        2        2          2           ⎛   4       3       2          ⎞                2           2
x  - 259⋅x  + 307⋅x + 147⎠ - 828⋅(1 - 3⋅x) ⋅(-3⋅x - 1) ⋅(x - 3) ⋅(x - 1) ⋅(3⋅x - 1) ⋅(3⋅x + 1)⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠ + 648⋅(1 - 3⋅x) ⋅(-3⋅x - 1) 
                                                                                                                                                            
                                                                                                                                                            

                                                                                                                                                            
        2        2          3 ⎛   4       3       2          ⎞               2           2        2          2          3 ⎛   4       3       2          ⎞  
⋅(x - 3) ⋅(x - 1) ⋅(3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠ - 84⋅(1 - 3⋅x) ⋅(-3⋅x - 1) ⋅(x - 3) ⋅(3⋅x - 1) ⋅(3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠ +
                                                                                                                                                            
                                                                                                                                                            

                                                                                                                                                            
             2           2                2          2          3 ⎛   4       3       2          ⎞            2           2        2          2          2  
 36⋅(1 - 3⋅x) ⋅(-3⋅x - 1) ⋅(x - 3)⋅(x - 1) ⋅(3⋅x - 1) ⋅(3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠ - (1 - 3⋅x) ⋅(-3⋅x - 1) ⋅(x - 1) ⋅(3⋅x - 2) ⋅(3⋅x - 1) ⋅(
                                                                                                                                                            
                                                                                                                                                            

                                                                                                                                                            
        3 ⎛   4       3       2          ⎞               2           2        2                    2          3 ⎛   4       3       2          ⎞            
3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠ - 12⋅(1 - 3⋅x) ⋅(-3⋅x - 1) ⋅(x - 1) ⋅(3⋅x - 2)⋅(3⋅x - 1) ⋅(3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠ + 84⋅(1 - 3
                                                                                                                                                            
                                                                                                                                                            

                                                                                                                                                            
   2           2        2          2          3 ⎛   4       3       2          ⎞            2           2           ⎛   4       3       2          ⎞        
⋅x) ⋅(-3⋅x - 1) ⋅(x - 1) ⋅(3⋅x - 1) ⋅(3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠ + (1 - 3⋅x) ⋅(-3⋅x - 1) ⋅(3⋅x + 1)⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠⋅((x - 1
                                                                                                                                                            
                                                                                                                                                            

                                                                                                                                                            
                                                                                                                                                            
)⋅(3⋅x - 1)⋅(3⋅x + 1)⋅(5⋅x - 15) + (x - 1)⋅(3⋅x - 1)⋅(23⋅x - 69) - (x - 1)⋅(3⋅x + 1)⋅(18⋅x - 54) + (3⋅x - 1)⋅(3⋅x + 1)⋅(7⋅x - 21) - (3⋅x - 1)⋅(3⋅x + 1)⋅(7⋅x
                                                                                                                                                            
                                                                                                                                                            

                                                                                                                                                            
      2                                          2                    2          3 ⎛   4       3       2          ⎞                             3        2  
 - 7))  - (72 - 216⋅x)⋅(-3⋅x - 1)⋅(x - 3)⋅(x - 1) ⋅(3⋅x - 2)⋅(3⋅x - 1) ⋅(3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠ - (2592 - 7776⋅x)⋅(-3⋅x - 1) ⋅(x - 3) ⋅(
                                                                                                                                                            
                                                                                                                                                            

                                                                                                                                         
      2          2 ⎛   4       3       2          ⎞               2        2          2          3 ⎛   4       3       2          ⎞⎞     
x - 1) ⋅(3⋅x - 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠ - 3888⋅(x - 3) ⋅(x - 1) ⋅(3⋅x - 1) ⋅(3⋅x + 1) ⋅⎝3⋅x  - 14⋅x  + 16⋅x  - 2⋅x - 3⎠⎠⋅p(x)
```
Now `psol = 1` and we do:
```python
cf = auxeq.subs(p(x), psol).doit().expand()
```
Before calling `expand` we have:
```python
(Pdb) p auxeq.subs(p(x), psol).doit()
-4*(1 - 3*x)**3*(-3*x - 1)*(x - 3)**2*(x - 1)**2*(3*x - 1)**2*(3*x + 1)**3*(12*x**4 - 67*x**3 - 259*x**2 + 307*x + 147) - 828*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)**2*(x - 1)**2*(3*x - 1)**2*(3*x + 1)*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) + 648*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)**2*(x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - 84*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) + 36*(1 - 3*x)**2*(-3*x - 1)**2*(x - 3)*(x - 1)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - (1 - 3*x)**2*(-3*x - 1)**2*(x - 1)**2*(3*x - 2)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - 12*(1 - 3*x)**2*(-3*x - 1)**2*(x - 1)**2*(3*x - 2)*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) + 84*(1 - 3*x)**2*(-3*x - 1)**2*(x - 1)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) + (1 - 3*x)**2*(-3*x - 1)**2*(3*x + 1)*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3)*(5*(x - 3)*(x - 1)*(3*x - 1)*(3*x + 1) + 23*(x - 3)*(x - 1)*(3*x - 1) - 18*(x - 3)*(x - 1)*(3*x + 1) + 7*(x - 3)*(3*x - 1)*(3*x + 1) - 7*(x - 1)*(3*x - 1)*(3*x + 1))**2 - 2592*(1 - 3*x)*(-3*x - 1)**3*(x - 3)**2*(x - 1)**2*(3*x - 1)**2*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - 72*(1 - 3*x)*(-3*x - 1)*(x - 3)*(x - 1)**2*(3*x - 2)*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3) - 3888*(x - 3)**2*(x - 1)**2*(3*x - 1)**2*(3*x + 1)**3*(3*x**4 - 14*x**3 + 16*x**2 - 2*x - 3)
```
After expanding we have
```python
(Pdb) p auxeq.subs(p(x), psol).doit().expand()
0
```
It is generally more efficient for this kind of thing to use `Poly` rather than `Expr`. With `Poly` everything is automatically expanded and the expansion of powers and products can be computed more efficiently. Example:
```python
In [49]: p = random_poly(x, 20, 0, 10)

In [50]: p2 = Poly(p)

In [51]: p
Out[51]: 
   20      19      18      17      16      15       14      13      12      11      10      9      8      7      6      5      4      3      2      
7⋅x   + 3⋅x   + 8⋅x   + 7⋅x   + 2⋅x   + 5⋅x   + 10⋅x   + 2⋅x   + 8⋅x   + 9⋅x   + 3⋅x   + 2⋅x  + 4⋅x  + 7⋅x  + 7⋅x  + 4⋅x  + 3⋅x  + 4⋅x  + 2⋅x  + 5⋅x

In [52]: p2
Out[52]: 
Poly(7*x**20 + 3*x**19 + 8*x**18 + 7*x**17 + 2*x**16 + 5*x**15 + 10*x**14 + 2*x**13 + 8*x**12 + 9*x**11 + 3*x**10 + 2*x**9 + 4*x**8 + 7*x**7 + 7*x**6 + 4*x*
*5 + 3*x**4 + 4*x**3 + 2*x**2 + 5*x, x, domain='ZZ')

In [53]: %time ok = expand(p**2)
CPU times: user 69.5 ms, sys: 719 µs, total: 70.2 ms
Wall time: 72 ms

In [54]: %time ok2 = p2**2
CPU times: user 424 µs, sys: 0 ns, total: 424 µs
Wall time: 433 µs

In [55]: ok2.as_expr() == ok
Out[55]: True
```
That's a 100x speedup but the difference can be more e.g. for `p**5` it's 1000x faster:
```python
In [1]: p = random_poly(x, 20, 0, 10)

In [2]: p2 = Poly(p)

In [3]: p
Out[3]: 
   20       19      18      17      16      15    14      13      12      11      9      8      6      5      4      3    2      
2⋅x   + 10⋅x   + 7⋅x   + 4⋅x   + 4⋅x   + 9⋅x   + x   + 6⋅x   + 9⋅x   + 4⋅x   + 8⋅x  + 8⋅x  + 3⋅x  + 7⋅x  + 5⋅x  + 9⋅x  + x  + 6⋅x

In [4]: %time ok = expand(p**5)
CPU times: user 3.84 s, sys: 20.8 ms, total: 3.86 s
Wall time: 3.87 s

In [5]: %time ok2 = p2**5
CPU times: user 2.34 ms, sys: 25 µs, total: 2.36 ms
Wall time: 2.38 ms
```
The thing about using `Poly` for speed though is that you really need to use it everywhere. The cost of converting backwards and forwards between Poly and Expr will otherwise dominate your calculation.